+++ addon/components/flexberry-simpleolv.js
@@ -1758,7 +1758,7 @@ ErrorableControllerMixin, {
 
     this._deleteHasManyRelationships(record, immediately).then(() => immediately ? record.destroyRecord() : record.deleteRecord()).catch((reason) => {
       this.rejectError(reason, `Unable to delete a record: ${record.toString()}.`);
-      record.rollbackAttributes();
+      record.rollbackAll();
     });
 
     let componentName = this.get('componentName');