+++ CHANGELOG.md
@@ -3,9 +3,14 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.7.0-beta.21] - 2016-10-16
 ### Added
 * `skipTransition` parameter to `delete` method (first parameter of method) of `edit-form` controller. Setting to true allows to skip technological call of `transitionToRoute` method.
 * `skipTransition` parameter to `saveAndClose`, `close` and `delete` actions (as first parameter) of `edit-form` and `detail-edit-form` controller. Setting to true allows to skip technological call of `transitionToRoute` method.
+
+### Changed
+* Updated dependency on `ember-flexberry-data` addon to v0.6.2-beta.23 version.
 
 ### Fixed
 * Detail models delete logic in offline mode.